src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/pseudo-host-parameter-matches-shadow-host-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Test that :host(selector) can match a shadow host when the host matches the selector: http://crbug.com/313935.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
